+++ another-file.js
@@ -1,7 +1,5 @@
-var myObject = {
+export const myObject = {
   stuff: 'some stuff here',
   otherStuff: 'some other stuff there',
   hellaStuff: 'hella stuff everywhere'
 }
-
-module.exports = myObject;