+++ grab-all-my-code-from-this-file.js
@@ -1,7 +1,16 @@
-document.write('Is this thing on?');
-var stuff = require('./another-file.js');
-console.dir(stuff);
-var element = document.createElement('h1');
-element.textContent = 'We\'re Webpacking!!!';
+import { myObject } from './another-file.js';
 
-document.body.appendChild(element);
+const stuffFunc = (thingy) => {
+  document.write('Is this thing on?');
+  console.dir(thingy);
+}
+
+stuffFunc(myObject);
+
+const testFunc = () => {
+  const element = document.createElement('h1');
+  element.textContent = 'We\'re Webpacking!!!';
+  document.body.appendChild(element);
+}
+
+testFunc();