+++ webpack.config.js
@@ -3,5 +3,10 @@ module.exports = {
   entry: './grab-all-my-code-from-this-file.js',
   output: {
     filename: './concat-and-put-all-that-stuff-here.js'
+  },
+  module: {
+    loaders: [
+      { test: /\.js$/, loader: 'babel', exclude: /node_modules/ }
+    ]
   }
 }